Move transformation methods to recoordinate_dialog.py
